fix: move asset facts service into assets package
